+++ changelog.md
@@ -1,5 +1,9 @@
 # Payplug module changelog
 
+## Version 4.16.0
+- Refactoring :
+  - [PRE-1999](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2615): Clean obsolete hook
+  
 ## Version 4.15.0
 - Feature :
   - [PRE-2673](https://gitlab.com/dalenys/public/ecommerce/prestashop_v4_17_8/-/merge_requests/2594): Allow refund for PPRO payment resource